Bring back support for voxel types with 8-byte quads

+++ src/main.cpp
@@ -231,7 +231,7 @@ void create_chunk() {
     glBufferSubData(
       GL_SHADER_STORAGE_BUFFER,
       0,
-      meshData.vertexCount * sizeof(uint32_t),
+      meshData.vertexCount * sizeof(uint64_t),
       meshData.vertices->data()
     );
   }
@@ -285,7 +285,7 @@ int main(int argc, char* argv[]) {
   meshData.opaqueMask = new uint64_t[CS_P2] { 0 };
   meshData.faceMasks = new uint64_t[CS_2 * 6] { 0 };
 
-  meshData.vertices = new std::vector<uint32_t>(10000);
+  meshData.vertices = new std::vector<uint64_t>(10000);
   meshData.maxVertices = 10000;
 
   create_chunk();
